chore: split functions into different file
